extra queries added, ingredients not outputting
