+++ Frontend/results.html
@@ -13,19 +13,6 @@
             </div>
             <button id="back" style="width: 100px; height: 30px; margin-top: 10px;" onclick="window.location.href='index.html';">Back</button>
         </div>
-        <script>
-            // Retrieve the query result from localStorage
-            const queryResult = JSON.parse(localStorage.getItem('queryResult'));
-    
-            // Display the result
-            const resultsDiv = document.getElementById('results');
-    if (queryResult && queryResult.length > 0) {
-        const data = document.createElement('p');
-        data.textContent = JSON.stringify(queryResult, null, 2); // Pretty-print JSON
-        resultsDiv.appendChild(data);
-    } else {
-        resultsDiv.textContent = 'No results found.';
-    }
-        </script>
+        <script src="results.js"></script>
     </body>
 </html>
